+++ platform/src/sdk/sdk.ts
@@ -63,9 +63,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "2.1.0";
+    sdkVersion = "2.2.0";
     genVersion = "2.223.0";
-    userAgent = "speakeasy-sdk/typescript 2.1.0 2.223.0 3.0.0 @codat/platform";
+    userAgent = "speakeasy-sdk/typescript 2.2.0 2.223.0 3.0.0 @codat/platform";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);